serializer_server/requirements.txt
avro==1.11.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dict2xml==1.7.3
dicttoxml==1.7.16
fastavro==1.7.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
xmltodict==0.13.0
